Finish writing documentation
