+++ src/components/Message/docs/Timestamp.md
@@ -1,3 +1,19 @@
 # Timestamp
 
 A Timestamp component renders the UI for the time and read status of a [Message](./Message.md).
+
+
+## Example
+
+```jsx
+<Timestamp timestamp='9:41am' read />
+```
+
+
+## Props
+
+| Prop | Type | Description |
+| --- | --- | --- |
+| className | string | Custom class names to be added to the component. |
+| read | boolean | Determines if the Message is read. |
+| timestamp | string | Timestamp for the Message. |
