+++ stories/Message.js
@@ -134,6 +134,15 @@ stories.add('content', () => (
   </Message>
 ))
 
+stories.add('media', () => (
+  <Message from avatar={<Avatar name='Arctic Puffin' />}>
+    <Message.Chat read timestamp='9:41am'>
+      Hey Buddy!
+    </Message.Chat>
+    <Message.Media imageUrl='https://img.buzzfeed.com/buzzfeed-static/static/2014-12/5/11/enhanced/webdr06/longform-original-7538-1417798667-22.jpg?downsize=715:*&output-format=auto&output-quality=auto' caption='image.jpg' timestamp='9:41am' read />
+  </Message>
+))
+
 stories.add('note', () => (
   <Message to avatar={<Avatar name='Buddy' />}>
     <Message.Chat isNote read timestamp='9:41am'>
